# Optimise sync of scraping targets