@bboreham 2023-03-07
Rebased on master and added another optimisation, which is a breaking change to the interface of `TargetsFromGroup`.
Will update benchmarks shortly.