@roidelapluie 2023-03-08
Thank you! I like how you used a for loop with `Get` in `resolveConflictingExposedLabels`.